refactor(l5): move distill-window and distiller from src/ into server/domain (issue 864)
L5 of the server-hardening ladder: cut server/ free of src/. Two source
modules move, their old src/ paths become shims so every legacy importer
keeps its line untouched.

Moves:
- src/distill-window.ts -> server/domain/distill-window.ts
- src/distiller.ts      -> server/domain/distiller.ts

Both moved files now meet the server/ standard whole, since the pre-commit
gate lints staged files entire:

- claimDistillBatch was 117 code lines at complexity 11. Its bind-parameter
  and SQL-fragment assembly is hoisted into module-scope buildClaimBindings,
  and its row decoding plus due-set scan into toDistillTurn / readClaimRows.
  The placeholder numbering and the fragments naming those placeholders are
  now produced in one place, which is the pairing that a hand-written $n in
  a body that also pushes params gets wrong.
- Indexed reads that used non-null assertions now read the element and take
  the same branch they took before when it is undefined: buildUnits skips a
  missing current turn and stops the backward context walk, and
  nearestSpeechContext skips. readClaimRows iterates rows by value instead
  of indexing two arrays in parallel.

No behavior change. Their imports of other src/ modules are rewritten to
../../src/ per M3; those files belong to later lanes.

Tests left in src/ (retire or move at L6): src/distill-window.test.ts and
src/distiller.test.ts stay byte-identical at their src/ paths and exercise
the moved code through the shims. Neither is staged.

Closure of src/ under server/main.ts: 50 -> 48.

Done-means: scripts/done-means/864-moved-out-of-src.sh src/distill-window.ts src/distiller.ts
